convert to markdown and get ast
add style to renderer

add new line to block elements
